feat: EventManager 클래스를 추가하고 FaceDetectionSDK에서 이벤트 관리 로직을 통합하여 콜백 처리 방식을 개선함
